+++ SIL.Media/FFmpegRunner.cs
@@ -169,38 +169,6 @@ namespace SIL.Media
 		private static ExecutionResult NoFFmpeg =>
 			new ExecutionResult { StandardError = "Could not locate FFmpeg" };
 
-		///<summary>
-		/// Returns false if it can't find ffmpeg
-		///</summary>
-		private static bool HaveValidFFmpegOnPath
-		{
-			get
-			{
-				if (Platform.IsWindows)
-				{
-					if (LocateFFmpeg() != null)
-						return true;
-				}
-
-				//see if there is one on the %path%
-
-				var p = new Process();
-				p.StartInfo.FileName = "ffmpeg";
-				p.StartInfo.RedirectStandardError = true;
-				p.StartInfo.UseShellExecute = false;
-				p.StartInfo.CreateNoWindow = true;
-				try
-				{
-					p.Start();
-				}
-				catch (Exception)
-				{
-					return false;
-				}
-				return true;
-			}
-		}
-
 		/// <summary>
 		/// Extracts the audio from a video. Note, it will fail if the file exists, so the client
 		/// is responsible for verifying with the user and deleting the file before calling this.
@@ -212,7 +180,7 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult ExtractMp3Audio(string inputPath, string outputPath, int channels, IProgress progress)
 		{
-			if (LocateFFmpeg() == null)
+			if (LocateAndRememberFFmpeg() == null)
 				return NoFFmpeg;
 
 			var arguments = $"-i \"{inputPath}\" -vn {mp3LameCodecArg} -ac {channels} \"{outputPath}\"";
@@ -247,7 +215,7 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult ExtractOggAudio(string inputPath, string outputPath, int channels, IProgress progress)
 		{
-			if (LocateFFmpeg() == null)
+			if (LocateAndRememberFFmpeg() == null)
 				return NoFFmpeg;
 
 			var arguments = $"-i \"{inputPath}\" -vn -acodec vorbis -ac {channels} \"{outputPath}\"";
@@ -328,7 +296,7 @@ namespace SIL.Media
 		private static ExecutionResult ExtractAudio(string inputPath, string outputPath,
 			string audioCodec, int sampleRate, int channels, IProgress progress)
 		{
-			if (LocateFFmpeg() == null)
+			if (LocateAndRememberFFmpeg() == null)
 				return NoFFmpeg;
 
 			var sampleRateArg = "";
@@ -371,7 +339,7 @@ namespace SIL.Media
 		public static ExecutionResult ChangeNumberOfAudioChannels(string inputPath,
 			string outputPath, int channels, IProgress progress)
 		{
-			if (LocateFFmpeg() == null)
+			if (LocateAndRememberFFmpeg() == null)
 				return NoFFmpeg;
 
 			var arguments = $"-i \"{inputPath}\" -vn -ac {channels} \"{outputPath}\"";